Compile linker and loader source files with no warnings using -Wall

+++ source/linker/kind.h
@@ -30,7 +30,7 @@ extern void WriteKinds();
 extern void WriteGKinds(int fd);
 extern void WriteLKinds(int fd);
 
-extern int CheckKindArity(KindInd i);
+extern Byte CheckKindArity(KindInd i);
 
 extern TwoBytes PackKindInd(KindInd ind);
 extern KindInd UnPackKindInd(TwoBytes ind);
